Debugging and pyc file update

+++ MedAI_functions.py
@@ -144,7 +144,7 @@ class train_test_split():
             self.test_size = 0.25
             # Compute train and test sizes
         if isinstance(self.test_size, float):
-            self._n_test = np.floor(self.test_size * self._n_samples)
+            self._n_test = int(np.floor(self.test_size * self._n_samples))
         elif isinstance(self.test_size, int):
             self._n_test = self.test_size
         else:
@@ -153,7 +153,7 @@ class train_test_split():
         if self.train_size is None:
             self._n_train = self._n_samples - self._n_test
         elif isinstance(self.train_size, float):
-            self._n_train = np.floor(self.train_size * self._n_samples)
+            self._n_train = int(np.floor(self.train_size * self._n_samples))
         elif isinstance(self.train_size, int):
             self._n_train = self.train_size
         else:
